feat: 게임 종료 시 미체결, 매도 하지 않은 매수거라 종가로 처리 로직 추가

+++ src/main/java/io/cavia/trader/module/game/repository/GameMapper.java
@@ -16,7 +16,7 @@ public interface GameMapper {
     Long findLastGameId();
     GameParticipation findLastGameParticipationByMemberId(Long memberId);
     void updateCashAndTotalScoreById(@Param("id") long id,
-                                     @Param("cash") int cash,
+                                     @Param("cash") long cash,
                                      @Param("totalScore") int totalScore
     );
 }
